+++ CHANGES.rst
@@ -105,3 +105,4 @@ v1.0.15
 
 v1.0.16[unreleased]
 -----------
+* Add name filtering to users